fix on `eval` with not exsisting file.
`eval` can specify unexisting file so it shouldn't read.
Without `eval` the file can be removed just after compiling,
the error should not be propagated.
If given file is existing file for eval context, it should not
override the source code.

The issue is pointed at https://github.com/ruby/debug/pull/10 by
casperisfine and this patc contains a part of his patch.