Make sure duplicated breakpoint is disabled
The debugger currently shows a warning message for duplicated
breakpoints and doesn't add it to the breakpoints list. But line or
catch breakpoints would still be activated and trigger a stop.

To fix this issue, we need to disable duplicated breakpoints. However,
`binding.bp` also uses line breakpoints with a fixed internal location,
simply excluding duplicated breakpoints will break it.

So the solution requires 2 checks:

1. Check if the breakpoint is duplicated by key.
2. Check if it's a normal breakpoint (e.g. added by the `break` command)
   instead of a oneshot (binding.bp or DEBUGGER__.console) breakpoint.

If both are satisfied, it should be disabled.