Only check line numbers after the test is finished
Users should be able to freely change their test script without the
interruption of the line numbers until the script is confirmed working.

Only then should the test framework check the script's line numbers with
a suggested change (if it doesn't have them yet).

This change will disable line number check on the test framework's
tests because most of those cases are designed to fail. But the number
check wasn't design for those tests anyway.