Allow accessing CatchBreakpoint's exception with _ex_ expression
Currently, there's no way to access the raised exception inside a
CatchBreakpoint, which is not convenient for debugging. So this commit
allows accessing the exception with the `_ex_` expression (similar to
Pry).

But because the breakpoint can stop at both Ruby method frame (has
binding) and C method frame (doesn't have binding), we can't assign the
exception to a universally accessible container.

My implementation hijacks the user expression without actually evaluate
it in the current scope. The drawback is that this `_ex_` expression
won't show up in variables or methods list. I think this isn't a big
problem as the exception isn't really stored in the program without
debugger.