Expand explanation of usage in source code
When I read the documentation, I thought it was saying I could use debug with `binding.irb` and `binding.pry` because of the wording that was chosen. I hope this change helps others to understand that the reference to those methods are simply examples to help readers understand how to use `binding.break`.