Fix the current 'bundle exec' skipping approach
When running `rdbg` executable with `bundle exec ...` commands, like

```
rdbg -c bundle exec cmd
```

we want to prevent the debugger from stopping inside the bundler. So the
debugger currently skips loading `session.rb` if the command contains
`bundle exec` with the line:

```ruby
// session.rb
return if $0.end_with?('bin/bundle') && ARGV.first == 'exec'
```

This approach assumes that `bundle exec cmd` will be executed as 2
separate commands:

- `bundle exec cmd`
- `cmd`

In those cases, `session.rb` will not be loaded in the `bundle exec cmd` run
but will be loaded when `cmd` is executed.

Which is true for:

- `bundle exec rails c`
  - `bundle exec rails c` - skipped
  - `rails c` - loaded
- `bundle exec ruby target.rb`
  - `bundle exec ruby target.rb` - skipped
  - `target.rb` - loaded

However, for commands like

- `bundle exec rspec _spec_`
- `bundle exec rake -T`
- `bundle exec irb`

they will be directly executed as `bundle exec ...`, so the `session.rb`
will be skipped and never loaded. This would then cause `DEBUGGER__` to be undefined.

(The cause of the difference between the 2 groups of commands is not clear
yet.)

This commit changes the approach from "skipping session.rb" to "not
setting initial breakpoint inside bin/bundle".