skip loading on `bundle exec` 2nd approach
`bundle exec` doesn't use `exec` if the loading command is
written in Ruby. To check this kind of invocation, obnserving
`$0` modification and setup debugger.

`RUBY_DEBUG_INITIAL_SUSPEND_PATH` envval is provided for the this
purpose and maybe we shouldn't document it if we can find general
usage of it.

Fix #267
See #269