Pin objects that are defined in Ruby.
`DEBUGGER__` and `FrameInfo` are defined in Ruby, so they are not
automatically pinned.  `rb_mDebugger` and `rb_cFrameInfo` are static
global variables that hold references to these objects.  If compaction
happens, these static references can go bad.  To fix this, we register
the two references as "mark objects" so that the GC will automatically
pin the references.

Fixes: #272