add `lib/debug/prelude.rb`
`RUBYOPT='-r (path/to/)lib/debug/prelude.rb'` enables to use
`debugger` method without `require 'debug'`.

`rdbg --util=init` shows what should be in `.bash_profile` and so on

```
$ rdbg --util=init

eval "$(rdbg init -)"
```

and `rdbg --util=init -` shows the RUBYOPT with (fullpath)/debug/prelude.

```
$ rdbg --util=init -
export RUBYOPT="-r /mnt/c/ko1/src/rb/ruby-debug/lib/debug/prelude $(RUBYOPT)"
```

Note that this feature is experimetanl.
It enables debugger just after `debugger` is called, so the tracking
feature should not be enough for example the laoded source code
before the loading debugger is not stored. `~/.rdbgrc` is not loaded
at the beggining of the program, but is loaded at using `debugger` method.
Please use `rdbg` or require `debug` if you want to enable full features.

On the other words, if you want to use it casually without full feature,
for example you only want to stop and check the program state with `debugger`
method, you only need it.

If you are so lazy, `RUBY_DEBUG_B` envval is defined, `Kernel#b` is
defined as an alias of `debugger` method.