support multi-process debugger mode
This commit allows to debug multi-process by forking.
If one process waiting for the input for REPL, all processes
(called process group in the source code) waits for entering
new interactive session (called subsession in the source code).
In other words, the process group has a global lock and a process
needs to acquire the lock to enter an subsession.

`parent_on_fork` configuration becomes obsolete (will be removed
soon) and `fork_mode` is introduced.

On the REPL, you will see PID and `$0` information on the prompt
on multi-process mode:

```
[1, 6] in target.rb
     1| fork{
     2|   $0 = 'child_process'
=>   3|   binding.b
     4| }
     5|
     6| __END__
=>#0    block in <main> at target.rb:3
  #1    [C] Kernel#fork at /mnt/c/ko1/src/rb/ruby-debug/lib/debug/session.rb:2037
  # and 1 frames (use `bt' command for all frames)
(rdbg@child_process#15907)
```