DAP: source on not local fs
When debugging on the other machine (non-local fs) with TCP/IP,
the source should be sent each time.

This patch check the local filesystem or not and if it is not a
local filesystem, specify `sourceReference` (and VSCode send
`source` request to get a whole source code).

On the TCP/IP, debugger assumes it is non-local fs, but
`'localfs':true` in `launch.json` tells that VSCode and
debugger run on same machine (share the same local fs).