CheckBreakpoint should only stop at the first condition fulfillment line
When a condition is fulfilled (e.g. a >= 1), it usually remains true for
at least a while. However, under the current implementation, it'll stops
at every single line after the condition becomes true, which is
annoying.

So this PR introduces an extra state to make sure the BP doesn't stop
repeatedly after the condition is fulfilled.