DAP: need to wait for connection
At the breakpoint by `debugger()` method, the connection is
needed and wait for the connection if it is not connected.

With the fowllowing scenario, the debugger raises an exception
becuase the connection is kept (but disconnection violate this
assumption)

1. run the script includes `debugger` call
2. connect to DAP client
3. disconnect from DAP client
4. call `debugger` method

To solve this issue, it needs to wait for the connection.
`UI_ServerBase::sock` and `UI_ServerBase::readline` do it and we
can use them on DAP connection, so remove `UI_DAP::sock` and
`UI_DAP::readline`.