Avoid setting @tc in thread_begin event
This event will be triggered before the new thread even has a
ThreadClient. So in this case, the @tc will be overridden to just nil,
which is wrong.

However, we also don't want to just set @tc to the new thread's
newly created ThreadClient. This is because a thread's creation can
happen during a subsession, and promptly switching @tc to another
thread's ThreadClient will misfire thread suspension.