change result chars: 4096 -> 180
Long results on DAP was clipped with 4096 chars. This patch shorten
this limit to 180 chars. Compound data structure doesn't have any
issue because expanding the result shows more details with corresponding
values such as instance variables and so on.

The problem is String object because we can not know the whole string body
if it is > 180 chars. To see the whole body this patch provides `#dump`
special field to show the whole body of the string and you can copy&paste
the value.

This patch also introduce `::DEBUGGER__::NaiveString` class. Evaluation
results of instances of this class will not be clipped so if you need to
get whole body of String, please wrap the string with `NaiveString` class.

fix https://github.com/ruby/debug/issues/802