CDP: do not override `UI_ServerBase::readline`
Recently, random failures such as `Timeout::Error` and `Errno::ECONNREFUSED` have been occurring. The reason for the error in CDP test is that UI_ServerBase::readline is overridden in CDP. continue is executed as a debug command because UI_CDP::readline returns continue if @q_msg is nil. See ono-max#8 for information on how to reproduce the error.