Display an accurate message when the test fails
Currently, the message when the test fails is overridden by a test that checks if the remote debuggee is terminated. For example, when `assert_line_text` method is failed, the message should be `Expected to include ~ in ~`, but it will be overridden and the message will be `Expected the remote program to finish ~`. We should not assert in ensure block to prevent the problem.