wait for the first suspending
The debugger shows "wait for debugger connection" message when
it waits for the remote debugger connection because it breaks
at the beggining of the program.

Without this synchronization, the debugger can attach *before*
the first temporary breakpoint (at the beggining of the program)
and the debug scenario breaks (and test fails).

It will fix some test failures.