Improve assert_repl_result for CDP
The current implementation doesn't work when the result is a boolean type
because the helper calls `#inspect` on the values of `JAVASCRIPT_TYPE_TO_CLASS_MAPS`,
which is an array for boolean types.

Example failure:

```
<["[TrueClass, FalseClass]"]> was expected to include
<"FalseClass">.
```

Since `JAVASCRIPT_TYPE_TO_CLASS_MAPS` is never used in other places, we
can just store the String representation of the class name in the hash
and thus avoid calling `#inspect` on the values.