Set evaluation result to `_` local variable
This PR makes a change to the thread client to set the `_` local
variable to the result of the last evaluation similar to the behavior in
irb[^1].

The reason `_` doesn't work even when using the `irb` console is because
we leave the `irb` evaluation early if the input should be handled by
the debugger[^2]. Otherwise, we would end up in IRB::Context#evaluate`
and we would call `set_last_value`[^3].

There's a bit of a quirk here when using the `irb` console. I was hoping
in `thread_client` to evaluate the result then set the `_` local
variable and the `@last_value` ivar all in the same method, similar to how `irb`
does it[^1]. However, because we're using the `irb` console we use a new
`irb` `Workspace` which nils out `_` when initialized[^4]. To get around
that, we set the local variable just before evaluating.

[^1]:(https://github.com/ruby/irb/blob/d43c3d764ae439706aa1b26a3ec299cc45eaed5b/lib/irb/context.rb#L460-L465)

[^2]:(https://github.com/ruby/irb/blob/d43c3d764ae439706aa1b26a3ec299cc45eaed5b/lib/irb.rb#L195-L198)

[^3]:(https://github.com/ruby/irb/blob/d43c3d764ae439706aa1b26a3ec299cc45eaed5b/lib/irb/context.rb#L550-L558)

[^4]:(https://github.com/ruby/irb/blob/d43c3d764ae439706aa1b26a3ec299cc45eaed5b/lib/irb/workspace.rb#L81)

This PR was created by Zendesk's Ruby infrastructure team:

Co-authored-by: Benjamin Quorning <bquorning@zendesk.com>
Co-authored-by: Edyta Rozczypała <edyta.rozczypala@zendesk.com>
Co-authored-by: Jury Razumau <jury.razumau@zendesk.com>
Co-authored-by: Leonid Batizhevskii <leonid.batizhevskii@zendesk.com>
Co-authored-by: Luis Manotas <lgerman@gmail.com>
Co-authored-by: Thomas Countz <thomascountz@gmail.com>